refactor(inputs): change components location
